Handle bisection sketch
If a peer responded to our request with a bisection sketch,
attempt to decode it to find missing transactions by combining
it with the initial sketch.
If success, share/request missing transactions.
If failure, send all the transactions we had for the peer.
In case of a partial failure, apply this logic accordingly to the parts.