Prepare to responding to bisection requests
Add 2 variables tracking reconciliation state:
(1) recon set snapshot and (2) capacity snapshot.
(1) is used to store transactions arrived after we sent out
an initial sketch, but before the reconciliation is over, since
these transactions should not go into a bisection sketch.
(2) is used to set the capacity of the bisection sketch:
making it different from the initial sketch is either pointless or inefficient.