Process incoming reconciliation request
Upon receiving a reconciliation request, a node stores
it and schedules a response.
Do not respond to a request right away as it would enable
a DoS attack and allow monitoring of transaction a node has.
Instead, respond to all reconciliation requests at a same time
after a small delay.