Assign announcement protocol based on tx source
If a transaction came from an inbound, flood.
This would facilitate rapid relay across the backbone
of reachable nodes with limited bandwidth.

If a transaction came from an outbound, reconcile.
This would allow to conserve most of the bandwidth.

If a transaction originated locally, reconcile.
This would avoid making obvious which transactions
originated at non-reachable nodes.