Add both txid and wtxid to m_recently_announced_invs
It's possible that a parent will be scheduled to relay same time as child,
but a child arrives earlier. Then, peer receiving an orphan would trigger
requesting a parent with txid, which would result in a NOTFOUND.

This happens because filterInventoryKnown has both txid and wtxid,
and prevents adding a parent txid to m_recently_announced_invs
when a peer requests a child with GETDATA.

Fix this by adding both wtxid and txid to m_recently_announced_invs
at the same time we schedule them for announcement, to allow
requesting parents even though they were scheduled for relay
(and not relayed yet) based on wtxid.